+++ exonum-light-client/src/main/lombok/com/exonum/client/response/ServiceInfo.java
@@ -12,26 +12,28 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
 package com.exonum.client.response;
 
-/**
- * Consensus status of a particular node.
- */
-public enum ConsensusStatus {
+import com.exonum.binding.common.message.TransactionMessage;
+import lombok.Value;
+
+@Value
+public class ServiceInfo {
+
   /**
-   * Shows that consensus is active,
-   * i.e., it is enabled and the node has enough connected peers.
+   * Returns the name of the service instance. It serves as the primary identifier of this service
+   * in most operations. It is assigned by the network administrators.
    */
-  ACTIVE,
+  String name;
+
   /**
-   * Shows that consensus is enabled on the node.
+   * Returns the numeric id of the service instance. Exonum assigns it to the service
+   * on instantiation. It is mainly used to route the transaction messages belonging
+   * to this instance.
+   *
+   * @see TransactionMessage#getServiceId()
    */
-  ENABLED,
-  /**
-   * Shows that consensus is disabled on the node.
-   */
-  DISABLED
+  int id;
 }