+++ exonum-light-client/CHANGELOG.md
@@ -25,6 +25,13 @@ The new release of the light client brings support for dynamic services.
 
 ### Added
 - Java 13 support.
+- `ExonumClient#findServiceInfo(String)` to retrieve a service id by its
+  name and `ExonumClient#getServiceInfoList` to retrieve the list of all
+  started services - their names and ids. (#1247)
+
+### Changed
+- `TransactionResponse#getExecutionResult` now returns `ExecutionStatus`
+  instead of `TransactionResult`. (#1244)
 
 ## 0.4.0 — 2019-10-09
 
@@ -39,7 +46,7 @@ and Exonum Java 0.8.
 - `ExonumClient#getBlockByHeight` and `#getBlocks` to throw
   `IllegalArgumentException` when blocks with heights exceeding
   the current blockchain height are requested (#1137)
-- `Block` JSON representation to be compatible with the one used 
+- `Block` JSON representation to be compatible with the one used
   for blocks by the core. Applied `@SerializedName` annotation
   to all fields. (#1137)
 - Updated project dependencies to the newest versions.